+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - optimized graph component
+- one zoom for all graphs in not united mode
 
 ## [1.73.0]