+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - optimized graph component
 - one zoom for all graphs in not united mode
+- fixed filter range for points and zoom multiline
 
 ## [1.73.0]
 
@@ -19,7 +20,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - fixed the datepicker in the dataset popup
-- fixed filter range for points and zoom multiline
 - closing the settings popup by clicking outside the popup
 - the json primitives text area has theme
 - added check for length of array metrics in manual filter and fixed save filters and use in other components