+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - optimized graph component
 - one zoom for all graphs in not united mode
+- add details menu for not email fields
 
 ## [1.73.0]