+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - added the slider component
+- added the ability to use tokens in filters
 
 ### Fixed
 - optimized graph component