+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - added the slider component
 - added the ability to use tokens in filters
+- the ability to use a token in the name of a part of the filter
 
 ### Fixed
 - optimized graph component